moved functional arguement packing code into the main header file, consider this a finished version 1
